Extracting properties method in a separate module
